Include port default fec configuration to be included in ZTP configuration
When using in-band interfaces for ZTP, it is important that the default
configuration of a port like speed and fec setting is configured correctly
for the port link to come up. This allows ZTP to perform DHCP discovery and
thus provision the device.

The proposed changes allow the ZTP configuration template to include port
fec mode only if it is defined.

Signed-off-by: Rajendra Dendukuri <rajendra.dendukuri@broadcom.com>